Conflicts:
	src/DefaultFramebuffer.cpp
	src/Framebuffer.cpp
	src/MeshTools/Interleave.h
	src/Plugins/MagnumFontConverter/MagnumFontConverter.cpp
	src/Primitives/Crosshair.cpp
	src/Primitives/Line.cpp
	src/Primitives/Plane.cpp
	src/Primitives/Square.cpp
	src/SceneGraph/AbstractTransformation.h
	src/SceneGraph/Object.hpp
	src/Shaders/DistanceFieldVector.cpp
	src/Shaders/Flat.cpp
	src/Shaders/Vector.cpp
	src/Text/GlyphCache.cpp
	src/TextureTools/Test/AtlasTest.cpp

+++ src/SceneGraph/Test/TranslationTransformationTest.cpp
@@ -28,9 +28,6 @@
 #include "SceneGraph/TranslationTransformation.h"
 #include "SceneGraph/Scene.h"
 
-/* For Object2Di */
-#include "SceneGraph/Object.hpp"
-
 namespace Magnum { namespace SceneGraph { namespace Test {
 
 typedef Object<TranslationTransformation2D> Object2D;
@@ -134,7 +131,7 @@ void TranslationTransformationTest::translate() {
 }
 
 void TranslationTransformationTest::integral() {
-    typedef Object<TranslationTransformation<2, Float, Short>> Object2Di;
+    typedef Object<TranslationTransformation<2, Float, Int>> Object2Di;
 
     Object2Di o;
     o.translate({3, -7});
